requirements.txt
torch
transformers
gradio
pytesseract
Pillow
numpy
